refactor out serviceworker code to serviceworkermanager
